plug-ins/jpeg/jpeg.[ch] in the JPEG save dialog, removed the option "Force
2007-08-10  Raphael Quinet  <raphael@gimp.org>

	* plug-ins/jpeg/jpeg.[ch]
	* plug-ins/jpeg/jpeg-save.[ch]: in the JPEG save dialog, removed
	the option "Force baseline JPEG" and added the new option "Use
	custom quantization tables" to allow resaving a JPEG file with the
	same compression settings as the original file. This is useful if
	the quantization tables are different from those that can be
	created by the IJG JPEG library.

	* plug-ins/jpeg/jpeg-settings.[ch]: updated comments, changed the
	return type of jpeg_restore_original_tables() because the IJG JPEG
	library expects unsigned integers.

svn path=/trunk/; revision=23197

+++ plug-ins/jpeg/jpeg-settings.h
@@ -27,5 +27,5 @@ gboolean  jpeg_restore_original_settings (gint32    image_ID,
                                           gint     *subsmp,
                                           gint     *num_quant_tables);
 
-guint16 **jpeg_restore_original_tables   (gint32    image_ID,
+guint   **jpeg_restore_original_tables   (gint32    image_ID,
                                           gint      num_quant_tables);
